pyproject.toml
[project]
name = "frigbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.79.0",
    "discord-py>=2.7",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

typeCheckingMode = "basic"
